wellness_backend/requirements.txt
fastapi
uvicorn
langchain
langchain[openai]
langchain_core
langchain_community
langgraph
tavily-python
langchain-tavily
langgraph-checkpoint-sqlite
python-dotenv
langchain-community
langchain-openai
langsmith